Fix errors from ultimate PR
