+++ wbooks/package.json
@@ -26,8 +26,6 @@
     "react": "16.13.1",
     "react-native": "0.63.4",
     "react-native-gesture-handler": "^1.9.0",
-    "react-native-linear-gradient": "^2.6.2",
-    "react-native-media-query": "^1.0.12",
     "react-native-safe-area-context": "^3.0.0",
     "react-native-screens": "3.13.1",
     "redux-mock-store": "^1.5.4",
