+++ wbooks/src/constants/colors.ts
@@ -4,3 +4,5 @@ export const white = '#FFFFFF';
 export const pictonBlue = '#38AAEB';
 export const aliceBlue = '#EFF9FF';
 export const WildSand = '#F5F5F5';
+export const EasternBlue = '#1EACB7';
+export const Turquoise = '#38CCCE';
