+++ wbooks/src/interfaces/BookProps.ts
@@ -1,0 +1,7 @@
+export interface BookProps {
+  title: string;
+  author: string;
+  imageUrl?: any;
+  id: number;
+  year: string;
+}
